pyproject.toml
[tool.poetry]
name = "cashback-demo"
version = "0.1"
description = "FastAPI demo for a cashback system"
authors = ["Tales Carlos de Pádua <talescpadua@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
sqlalchemy = "^1.3.19"
uvicorn = "^0.11.8"
mypy = "^0.782"
email-validator = "^1.1.1"
alembic = "^1.4.2"
mysqlclient = "^2.0.1"
python-jose = "^3.2.0"
cryptography = "^3.1"
bcrypt = "^3.2.0"
passlib = "^1.7.2"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.782"
pytest = "^6.0.1"
pytest-cov = "^2.10.1"
flake8 = "^3.8.3"
factory-boy = "^3.0.1"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
